Version 2 (#5)
add make file
add linting and formatting with ruff/pre-commit
add templating for different workflows for different envs (with jinja)
eliminate chispa, now it's possible to assert dataframe equalities with PySpark.
add a mechanism to choose tasks to be executed on dev env
add a mechanism to disable tasks in prod env.
improve task parameters
add medallion architecture example and tests for it